modules/utils/requirements.txt
avro
bson
cbor2
dicttoxml
kombu
msgpack
xmltodict
pyyaml